requirements.txt
numpy   2.2.6
matplotlib  3.10.7
scipy   1.16.3
scikit-image    0.25.2
Pillow  12.0.0